Some additional changes to the model and pipeline

+++ TPS_may/reduce_mem.py
@@ -20,8 +20,10 @@ def reduce_memory_usage(df, verbose=True):
                 elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                     df[col] = df[col].astype(np.in64)
             else:
+                # Float-16 causing issues with .parquet filetype
                 if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
-                    df[col] = df[col].astype(np.float16)
+                    #df[col] = df[col].astype(np.float16)
+                    df[col] = df[col].astype(np.float32)
                 elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                     df[col] = df[col].astype(np.float32)
                 else:
